Simplify release update handling
